Delete only be the user who added the blog.

+++ part-4/controllers/blogRoute.js
@@ -3,21 +3,11 @@ const blogRouter = require('express').Router()
 const Blog = require('../models/blogSchema')
 const User = require('../models/userSchema')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer')) {
-    return authorization.substring(7)
-  }
-  return null
-}
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {username : 1, name: 1})
   response.json(blogs)
 })
-  
 
-//need to change to async await 
-// eslint-disable-next-line no-unused-vars
 blogRouter.post('/', async (request, response, next) => {
   const body = request.body
   const token = request.token
@@ -28,7 +18,6 @@ blogRouter.post('/', async (request, response, next) => {
   }
   const user = await User.findById(decodedToken.id)
 
-
   const blog = new Blog({
     title: body.title,
     url: body.url,
@@ -37,13 +26,13 @@ blogRouter.post('/', async (request, response, next) => {
     user: user._id
   })
   const savedBlog = await blog.save()
-  user.blogs = user.blogs.concat(savedBlog._id)
-  await user.save()
-  response.json(savedBlog)
-  // try {
-  // } catch (exception) {
-  //   next(exception)
-  // }
+  try {
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.json(savedBlog)
+  } catch (exception) {
+    next(exception)
+  }
 })
 
 blogRouter.get('/:id', async (request, response,next) => {
@@ -54,17 +43,24 @@ blogRouter.get('/:id', async (request, response,next) => {
     } else {
       response.status(400).end()
     }
-  } catch(error) { next(error)}
+  } catch(error) {next(error)}
 })
 
 
 blogRouter.delete('/:id', async (request, response, next) => {
-  try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
-  } catch (error) {
-    next(error)
-  } 
+  if (!request.token || !request.decodedToken) {
+    return response.status(401).json({ error: 'missing or invalid token' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  const userId = request.decodedToken.id
+
+  if (blog.user.toString() !== userId.toString()) {
+    response.status(400).end()
+  }
+
+  await Blog.findByIdAndRemove(request.params.id)
+  response.status(204).end()
 })
 
 blogRouter.put('/:id', (request, response, next) => {
